move common function to keeper_internal_test.go
