+++ client/x/evmengine/keeper/db_internal_test.go
@@ -3,38 +3,10 @@ package keeper
 import (
 	"testing"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	"github.com/ethereum/go-ethereum/beacon/engine"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
-
-	moduletestutil "github.com/piplabs/story/client/x/evmengine/testutil"
-	"github.com/piplabs/story/lib/ethclient/mock"
-
-	"go.uber.org/mock/gomock"
 )
-
-func createTestKeeper(t *testing.T) (sdk.Context, *Keeper) {
-	t.Helper()
-	cdc := getCodec(t)
-	txConfig := authtx.NewTxConfig(cdc, nil)
-
-	mockEngine, err := newMockEngineAPI(0)
-	require.NoError(t, err)
-
-	ctrl := gomock.NewController(t)
-	mockClient := mock.NewMockClient(ctrl)
-	ak := moduletestutil.NewMockAccountKeeper(ctrl)
-	esk := moduletestutil.NewMockEvmStakingKeeper(ctrl)
-	uk := moduletestutil.NewMockUpgradeKeeper(ctrl)
-	ctx, storeService := setupCtxStore(t, nil)
-	ctx = ctx.WithExecMode(sdk.ExecModeFinalize)
-	keeper, err := NewKeeper(cdc, storeService, &mockEngine, mockClient, txConfig, ak, esk, uk)
-	require.NoError(t, err)
-
-	return ctx, keeper
-}
 
 func TestKeeper_InsertGenesisHead(t *testing.T) {
 	t.Parallel()
